Fix breaking tests by moving the coordinator to __init__
