+++ homeassistant/components/nightscout/const.py
@@ -1,6 +1,8 @@
 """Constants for the Nightscout integration."""
 
 DOMAIN = "nightscout"
+COORDINATOR = "coordinator"
+API = "api"
 
 ATTR_DEVICE = "device"
 ATTR_DELTA = "delta"
